chore: Move sales channel tests
